LayoutTests/platform/mac-wk1/accessibility/custom-elements/liveregions-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL accessibilityController.accessibleElementById("custom-element") should be null (of type object). Was undefined (of type undefined).
PASS accessibilityController.accessibleElementById("custom-liveregion").boolAttributeValue("AXARIAAtomic") is true
PASS accessibilityController.accessibleElementById("custom-liveregion").boolAttributeValue("AXElementBusy") is true
PASS accessibilityController.accessibleElementById("custom-liveregion").stringAttributeValue("AXARIALive") is "polite"
PASS accessibilityController.accessibleElementById("custom-liveregion").stringAttributeValue("AXARIARelevant") is "additions"
PASS accessibilityController.accessibleElementById("text-liveregion").boolAttributeValue("AXARIAAtomic") is false
PASS accessibilityController.accessibleElementById("text-liveregion").boolAttributeValue("AXElementBusy") is false
PASS accessibilityController.accessibleElementById("text-liveregion").stringAttributeValue("AXARIALive") is "assertive"
PASS accessibilityController.accessibleElementById("text-liveregion").stringAttributeValue("AXARIARelevant") is "text"
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
